# docs/openapi.json
{
  "components": {
    "schemas": {
      "AdjacentRelationshipRead": {
        "additionalProperties": false,
        "properties": {
          "counterpart": {
            "$ref": "#/components/schemas/WorkPointer"
          },
          "direction": {
            "enum": [
              "incoming",
              "outgoing",
              "undirected"
            ],
            "title": "Direction",
            "type": "string"
          },
          "relationship": {
            "$ref": "#/components/schemas/RelationshipEdgeRead"
          },
          "relative_to_work_item_id": {
            "format": "uuid",
            "title": "Relative To Work Item Id",
            "type": "string"
          }
        },
        "required": [
          "relationship",
          "relative_to_work_item_id",
          "direction",
          "counterpart"
        ],
        "title": "AdjacentRelationshipRead",
        "type": "object"
      },
      "CanonicalWorkProjection": {
        "additionalProperties": false,
        "properties": {
          "canonical_work_item": {
            "$ref": "#/components/schemas/WorkIdentityPointer"
          },
          "direct_destination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkIdentityPointer"
              },
              {
                "type": "null"
              }
            ]
          },
          "duplicate_member_count": {
            "minimum": 0.0,
            "title": "Duplicate Member Count",
            "type": "integer"
          },
          "is_duplicate": {
            "title": "Is Duplicate",
            "type": "boolean"
          },
          "path": {
            "items": {
              "$ref": "#/components/schemas/WorkIdentityPointer"
            },
            "maxItems": 50,
            "title": "Path",
            "type": "array"
          }
        },
        "required": [
          "is_duplicate",
          "direct_destination",
          "canonical_work_item",
          "path",
          "duplicate_member_count"
        ],
        "title": "CanonicalWorkProjection",
        "type": "object"
      },
...
      "CheckpointRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "context",
              "progress",
              "completion"
            ],
            "title": "Kind",
            "type": "string"
          },
          "legacy_record_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Legacy Record Id"
          },
          "migration_origin": {
            "anyOf": [
              {
                "enum": [
                  "legacy-handoff-snapshot",
                  "legacy-comment"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Migration Origin"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "repository_branch": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Branch"
          },
          "source_client": {
            "title": "Source Client",
            "type": "string"
          },
          "source_metadata": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Source Metadata",
            "type": "object"
          },
          "source_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Model"
          },
          "source_session_id": {
            "title": "Source Session Id",
            "type": "string"
          },
          "source_session_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Session Url"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "verified_against": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verified Against"
          },
          "work_item_id": {
            "format": "uuid",
            "title": "Work Item Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "work_item_id",
          "kind",
          "prompt",
          "source_client",
          "source_session_id",
          "source_model",
          "source_session_url",
          "repository_branch",
          "verified_against",
          "tags",
          "source_metadata",
          "migration_origin",
          "legacy_record_id",
          "created_at"
        ],
        "title": "CheckpointRead",
        "type": "object"
      },
      "ClaimAndRecall": {
        "additionalProperties": false,
        "properties": {
          "context": {
            "$ref": "#/components/schemas/WorkContext"
          },
          "lease": {
            "$ref": "#/components/schemas/ClaimReceipt"
          }
        },
        "required": [
          "lease",
          "context"
        ],
        "title": "ClaimAndRecall",
        "type": "object"
      },
      "ClaimReceipt": {
        "additionalProperties": false,
        "properties": {
          "acquired_at": {
            "title": "Acquired At",
            "type": "string"
          },
          "claim_request_id": {
            "title": "Claim Request Id",
            "type": "string"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "string"
          },
          "holder_client": {
            "title": "Holder Client",
            "type": "string"
          },
          "holder_session_id": {
            "title": "Holder Session Id",
            "type": "string"
          },
          "lease_token": {
            "title": "Lease Token",
            "type": "string"
          },
          "renewed_at": {
            "title": "Renewed At",
            "type": "string"
          },
          "work_item_id": {
            "format": "uuid",
            "title": "Work Item Id",
            "type": "string"
          }
        },
        "required": [
          "holder_client",
          "holder_session_id",
          "acquired_at",
          "renewed_at",
          "expires_at",
          "work_item_id",
          "claim_request_id",
          "lease_token"
        ],
        "title": "ClaimReceipt",
        "type": "object"
      },
...
      "DuplicateMergeEligibility": {
        "additionalProperties": false,
        "properties": {
          "has_unresolved_gate": {
            "title": "Has Unresolved Gate",
            "type": "boolean"
          },
          "incident_blocks_count": {
            "minimum": 0.0,
            "title": "Incident Blocks Count",
            "type": "integer"
          },
          "incident_parent_child_count": {
            "minimum": 0.0,
            "title": "Incident Parent Child Count",
            "type": "integer"
          },
          "source_lease_state": {
            "enum": [
              "none",
              "expired",
              "active"
            ],
            "title": "Source Lease State",
            "type": "string"
          }
        },
        "required": [
          "incident_blocks_count",
          "incident_parent_child_count",
          "has_unresolved_gate",
          "source_lease_state"
        ],
        "title": "DuplicateMergeEligibility",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "LeaseReleaseCreate": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MutationActor"
              },
              {
                "type": "null"
              }
            ]
          },
          "client_operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional caller-generated UUID for durable replay of this mutation. Reuse it only with the exact same operation and semantic arguments after an unknown outcome.",
            "title": "Client Operation Id"
          },
          "lease_token": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Lease Token",
            "type": "string"
          }
        },
        "required": [
          "lease_token"
        ],
        "title": "LeaseReleaseCreate",
        "type": "object"
      },
      "LeaseTokenCreate": {
        "additionalProperties": false,
        "properties": {
          "lease_token": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Lease Token",
            "type": "string"
          }
        },
        "required": [
          "lease_token"
        ],
        "title": "LeaseTokenCreate",
        "type": "object"
      },
      "MergeReviewRevision": {
        "additionalProperties": false,
        "description": "Exact, commit-visible context reviewed before an irreversible merge.",
        "properties": {
          "context_checkpoint_id": {
            "format": "uuid",
            "title": "Context Checkpoint Id",
            "type": "string"
          },
          "work_event_count": {
            "minimum": 1.0,
            "title": "Work Event Count",
            "type": "integer"
          },
          "work_version": {
            "minimum": 1.0,
            "title": "Work Version",
            "type": "integer"
          }
        },
        "required": [
          "work_version",
          "context_checkpoint_id",
          "work_event_count"
        ],
        "title": "MergeReviewRevision",
        "type": "object"
      },
      "MutationActor": {
        "additionalProperties": false,
        "description": "Client-asserted mutation provenance; never an authenticated identity.",
        "properties": {
          "actor_client": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Actor Client",
            "type": "string"
          },
          "actor_model": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor Model"
          },
          "actor_session_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Actor Session Id",
            "type": "string"
          }
        },
        "required": [
          "actor_client",
          "actor_session_id"
        ],
        "title": "MutationActor",
        "type": "object"
      },
...
      "RelationshipCounts": {
        "additionalProperties": false,
        "properties": {
          "incoming": {
            "default": 0,
            "title": "Incoming",
            "type": "integer"
          },
          "outgoing": {
            "default": 0,
            "title": "Outgoing",
            "type": "integer"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "undirected": {
            "default": 0,
            "title": "Undirected",
            "type": "integer"
          }
        },
        "title": "RelationshipCounts",
        "type": "object"
      },
...
      "ReleaseResult": {
        "additionalProperties": false,
        "properties": {
          "released": {
            "title": "Released",
            "type": "boolean"
          },
          "work_item_id": {
            "format": "uuid",
            "title": "Work Item Id",
            "type": "string"
          }
        },
        "required": [
          "work_item_id",
          "released"
        ],
        "title": "ReleaseResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkClaimCreate": {
        "additionalProperties": false,
        "properties": {
          "claim_request_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Claim Request Id",
            "type": "string"
          },
          "holder_client": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Holder Client",
            "type": "string"
          },
          "holder_session_id": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Holder Session Id",
            "type": "string"
          }
        },
        "required": [
          "holder_client",
          "holder_session_id",
          "claim_request_id"
        ],
        "title": "WorkClaimCreate",
        "type": "object"
      },
...
      "WorkContext": {
        "additionalProperties": false,
        "properties": {
          "canonical": {
            "$ref": "#/components/schemas/CanonicalWorkProjection"
          },
          "checkpoint_total": {
            "description": "Every checkpoint on this work item, including those not returned here.",
            "title": "Checkpoint Total",
            "type": "integer"
          },
          "current_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CheckpointRead"
              },
              {
                "type": "null"
              }
            ],
            "description": "The newest context checkpoint, or null when it is the initial checkpoint. When current_context_is_initial is true, read initial_checkpoint instead; the body is never serialized twice."
          },
          "current_context_is_initial": {
            "description": "True when the newest context checkpoint is the initial checkpoint.",
            "title": "Current Context Is Initial",
            "type": "boolean"
          },
          "duplicate_member_total": {
            "minimum": 0.0,
            "title": "Duplicate Member Total",
            "type": "integer"
          },
          "duplicate_members": {
            "items": {
              "$ref": "#/components/schemas/WorkIdentityPointer"
            },
            "maxItems": 20,
            "title": "Duplicate Members",
            "type": "array"
          },
          "duplicate_merge_eligibility": {
            "$ref": "#/components/schemas/DuplicateMergeEligibility"
          },
          "event_total": {
            "title": "Event Total",
            "type": "integer"
          },
          "incoming_relationships": {
            "items": {
              "$ref": "#/components/schemas/AdjacentRelationshipRead"
            },
            "maxItems": 100,
            "title": "Incoming Relationships",
            "type": "array"
          },
          "initial_checkpoint": {
            "$ref": "#/components/schemas/CheckpointRead"
          },
          "merge_review_revision": {
            "$ref": "#/components/schemas/MergeReviewRevision"
          },
          "omitted_checkpoint_count": {
            "description": "checkpoint_total minus the distinct checkpoints in this payload. Page the remainder with list_checkpoints.",
            "title": "Omitted Checkpoint Count",
            "type": "integer"
          },
          "omitted_duplicate_member_count": {
            "minimum": 0.0,
            "title": "Omitted Duplicate Member Count",
            "type": "integer"
          },
          "omitted_event_count": {
            "title": "Omitted Event Count",
            "type": "integer"
          },
          "omitted_relationship_counts": {
            "$ref": "#/components/schemas/RelationshipCounts"
          },
          "omitted_resolved_gate_count": {
            "title": "Omitted Resolved Gate Count",
            "type": "integer"
          },
          "omitted_unresolved_gate_count": {
            "title": "Omitted Unresolved Gate Count",
            "type": "integer"
          },
          "outgoing_relationships": {
            "items": {
              "$ref": "#/components/schemas/AdjacentRelationshipRead"
            },
            "maxItems": 100,
            "title": "Outgoing Relationships",
            "type": "array"
          },
          "pre_phase5_history_may_be_incomplete": {
            "title": "Pre Phase5 History May Be Incomplete",
            "type": "boolean"
          },
          "readiness": {
            "$ref": "#/components/schemas/Readiness"
          },
          "recent_checkpoints": {
            "description": "Older checkpoints for context, newest first. Never repeats the checkpoint returned as initial_checkpoint or current_context.",
            "items": {
              "$ref": "#/components/schemas/CheckpointRead"
            },
            "title": "Recent Checkpoints",
            "type": "array"
          },
          "recent_events": {
            "items": {
              "$ref": "#/components/schemas/WorkEventRead"
            },
            "title": "Recent Events",
            "type": "array"
          },
          "recent_resolved_gates": {
            "items": {
              "$ref": "#/components/schemas/HumanGateRead"
            },
            "title": "Recent Resolved Gates",
            "type": "array"
          },
          "relationship_counts": {
            "$ref": "#/components/schemas/RelationshipCounts"
          },
          "resolved_gate_total": {
            "title": "Resolved Gate Total",
            "type": "integer"
          },
          "undirected_relationships": {
            "items": {
              "$ref": "#/components/schemas/AdjacentRelationshipRead"
            },
            "maxItems": 100,
            "title": "Undirected Relationships",
            "type": "array"
          },
          "unresolved_gate_total": {
            "title": "Unresolved Gate Total",
            "type": "integer"
          },
          "unresolved_gates": {
            "items": {
              "$ref": "#/components/schemas/HumanGateRead"
            },
            "title": "Unresolved Gates",
            "type": "array"
          },
          "work_item": {
            "$ref": "#/components/schemas/WorkItemRead"
          }
        },
        "required": [
          "work_item",
          "merge_review_revision",
          "canonical",
          "duplicate_members",
          "duplicate_member_total",
          "omitted_duplicate_member_count",
          "initial_checkpoint",
          "current_context",
          "current_context_is_initial",
          "recent_checkpoints",
          "checkpoint_total",
          "omitted_checkpoint_count",
          "readiness",
          "unresolved_gates",
          "unresolved_gate_total",
          "omitted_unresolved_gate_count",
          "recent_resolved_gates",
          "resolved_gate_total",
          "omitted_resolved_gate_count",
          "duplicate_merge_eligibility",
          "recent_events",
          "event_total",
          "omitted_event_count",
          "pre_phase5_history_may_be_incomplete"
        ],
        "title": "WorkContext",
        "type": "object"
      },
...
      "WorkIdentityPointer": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "deferred",
              "done",
              "wont-do",
              "promoted"
            ],
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status"
        ],
        "title": "WorkIdentityPointer",
        "type": "object"
      },
...
      "WorkItemRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "initial_checkpoint_id": {
            "format": "uuid",
            "title": "Initial Checkpoint Id",
            "type": "string"
          },
          "priority": {
            "title": "Priority",
            "type": "integer"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "deferred",
              "done",
              "wont-do",
              "promoted"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "id",
          "project_id",
          "title",
          "summary",
          "status",
          "priority",
          "initial_checkpoint_id",
          "version"
        ],
        "title": "WorkItemRead",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Durable project-scoped work with immutable agent checkpoints.",
    "title": "Mnemonic API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/claim": {
      "post": {
        "operationId": "claim_work_api_v1_projects__project_id__work_items__work_item_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkClaimCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Claim Work"
      }
    },
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/claim-and-recall": {
      "post": {
        "operationId": "claim_and_recall_api_v1_projects__project_id__work_items__work_item_id__claim_and_recall_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkClaimCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimAndRecall"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Claim And Recall"
      }
    },
...
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/release-claim": {
      "post": {
        "operationId": "release_claim_api_v1_projects__project_id__work_items__work_item_id__release_claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseReleaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReleaseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Release Claim"
      }
    },
    "/api/v1/projects/{project_id}/work-items/{work_item_id}/renew-claim": {
      "post": {
        "operationId": "renew_claim_api_v1_projects__project_id__work_items__work_item_id__renew_claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "work_item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Work Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeaseTokenCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Renew Claim"
      }
    }
  },
...
}